updates for monday and such
