+++ src/lib/custom-events.js
@@ -1,0 +1,6 @@
+export default {
+    changeRecordColor: 'change-record-color',
+    changeRecordColorOut: 'change-record-color-out',
+    changeRecordColorInOut: 'change-record-color-in-out',
+    recordBackIn: 'record-back-in',
+}
