+++ src/lib/index.js
@@ -1,7 +1,7 @@
 // place files you want to import through the `$lib` alias in this folder.
 export const CURRENT_TEXTURE = 'current_texture';
 
-export const DATA_VERSION = "1.32122121111"
+export const DATA_VERSION = "1.321221211112"
 
 /** @type {Record<string, string>} */
 export const colors = {
